requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
matplotlib>=3.6.0
seaborn>=0.12.0
scikit-learn>=1.3.0
numpy>=1.24.0
transformers>=4.30.0
torch>=2.0.0
openai>=1.0.0
python-dateutil>=2.8.0
fpdf2>=2.7.0
python-dotenv>=1.0.0
networkx>=3.2.1
psutil>=5.9.8
reportlab>=4.2.2
accelerate>=0.33.0
bitsandbytes>=0.43.0 ; platform_system != 'Windows'
peft>=0.11.1
trl>=0.8.6
datasets>=2.20.0
sentencepiece>=0.2.0
pyyaml>=6.0.1
